Implemented CoverageTracker mixin to allow coverage tracking regardless of the selected base strategy.
