+++ src/strategies/mixins/FluentCoverageGuidanceMixin.ts
@@ -1,10 +1,4 @@
-import * as fs from 'fs'
-import * as utils from './utils'
-import * as schema from '@istanbuljs/schema'
-import * as libInstrument from 'istanbul-lib-instrument'
-
 import {performance} from 'perf_hooks'
-import {FluentCoverage} from '../FluentCoverage'
 import {MixinStrategy} from '../FluentStrategyTypes'
 import {FluentPick, WrapFluentPick} from '../../arbitraries'
 import {FluentStrategyInterface} from '../FluentStrategy'
@@ -13,68 +7,9 @@ export function CoverageGuidance<TBase extends MixinStrategy>(Base: TBase) {
   return class extends Base implements FluentStrategyInterface {
 
     /**
-     * Contains all the created names for the test methods.
-     */
-    private testMethodsNames: string[] = []
-
-    /**
      * Indicates whether the current test case collection inputs are the result of a mutation or not.
      */
     private testCaseCollectionMutationStatus = false
-
-    /**
-     * Instrumenter object.
-     */
-    private instrumenter = libInstrument.createInstrumenter({
-      parserPlugins: schema.defaults.nyc.parserPlugins.concat('typescript')
-    })
-
-    /**
-     * Coverage object responsible for computing and managing coverage
-     */
-    private coverageBuilder!: FluentCoverage
-
-    /**
-     * Function responsible for creating all the files needed so that coverage can be tracked.
-     */
-    protected coverageSetup() {
-      const coverageID = utils.generateUniqueMethodIdentifier()
-      const importInfo = utils.extractImports(this.configuration.importsPath, coverageID)
-      let sourceData: string = importInfo.header
-      const instrumentedFiles: string[] = importInfo.sourceFiles
-        .map(file => {
-          const fileArr = file.split('src/')
-          return fileArr[0] + 'src/.instrumented/' + coverageID + '/' + fileArr[1]
-        })
-
-      for (const file of instrumentedFiles) {
-        const sourceFile = file.split('.instrumented/' + coverageID + '/').join('') + '.ts'
-        utils.writeDataToFile(file + '.ts', this.instrumenter.instrumentSync(
-          fs.readFileSync(sourceFile).toString(), sourceFile))
-      }
-
-      for (const method of this.testMethods) {
-        this.testMethodsNames.push(utils.generateUniqueMethodIdentifier())
-        sourceData += 'export const ' + this.testMethodsNames[this.testMethodsNames.length - 1] + ' = '
-          + method.toString().replace(new RegExp(/[a-zA-Z]+_\d+\./gm), '') + '\n'
-      }
-
-      utils.writeDataToFile('src/.coverage/' + coverageID + '.ts', this.instrumenter.instrumentSync(
-        sourceData, 'src/.coverage/' + coverageID + '.ts'))
-
-      this.coverageBuilder = new FluentCoverage(coverageID + '.ts')
-    }
-
-    /**
-     * Resets the coverage global variable and removes all the files used for coverage purposes.
-     */
-    protected coverageTearDown() {
-      const paths = ['./src/.coverage', './src/.instrumented']
-      this.coverageBuilder.resetCoverage()
-
-      for (const path of paths)
-        utils.deleteFromFileSystem(path)
-    }
 
     /**
      * Generates each arbitrary seed collection, which is based on the arbitrary corner cases and extracted
@@ -102,7 +37,8 @@ export function CoverageGuidance<TBase extends MixinStrategy>(Base: TBase) {
     hasInput(): boolean {
       this.currTime = performance.now()
 
-      if (this.coverageBuilder.getTotalCoverage() >= this.configuration.coveragePercentage ||
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      if (this.getCoverageBuilder()!.getTotalCoverage() >= this.configuration.coveragePercentage ||
         this.configuration.timeout < this.currTime - (this.initTime ?? this.currTime)) return false
       else if (this.testCaseCollectionPick >= this.testCaseCollection.length) {
         for (const name in this.arbitraries) {
@@ -142,22 +78,17 @@ export function CoverageGuidance<TBase extends MixinStrategy>(Base: TBase) {
     handleResult(inputData: any[]) {
       inputData.forEach(data => {
         this.addTestCase(data)
-        this.coverageBuilder.compute(data)
+        this.getCoverageBuilder()?.compute(data)
       })
 
-      this.coverageBuilder.updateTotalCoverage()
+      this.getCoverageBuilder()?.updateTotalCoverage()
 
-      if (this.testCaseCollectionMutationStatus && this.coverageBuilder.compare())
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      if (this.testCaseCollectionMutationStatus && this.getCoverageBuilder()!.compare())
         for (const name in this.arbitraries)
           this.arbitraries[name].seedCollection = [... new Set(this.arbitraries[name].seedCollection
             .concat([this.currTestCase[name]]))]
     }
 
-    /**
-     * Returns the coverage builder coverage summary report.
-     */
-    getCoverage(): number {
-      return this.coverageBuilder.getTotalCoverage()
-    }
   }
 }
