+++ src/strategies/mixins/internal.ts
@@ -5,3 +5,4 @@ export * from './FluentDedupableMixin'
 export * from './FluentRandomMixin'
 export * from './FluentShrinkableMixin'
 export * from './FluentCoverageGuidanceMixin'
+export * from './FluentCoverageTracker'
